packages/datafrey-cli/pyproject.toml
[project]
name = "datafrey"
version = "0.1.0"
description = "CLI for managing Datafrey database connections and MCP servers"
license = "Apache-2.0"
requires-python = ">=3.13"
authors = [{name = "DataFrey", email = "slava@datafrey.ai"}]

dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "httpx>=0.27",
    "keyring>=25.0",
    "questionary>=2.1",
    "cryptography>=44.0",
    "PyJWT>=2.9",
    "datafrey-api",
]

[project.urls]
Homepage = "https://datafrey.ai"
Documentation = "https://docs.datafrey.ai"
Repository = "https://github.com/datafrey-ai/datafrey"

[project.scripts]
datafrey = "datafrey.cli.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
datafrey-api = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/datafrey"]
